+++ app/login/oauth2/code/google/page.tsx
@@ -2,7 +2,7 @@
 'use strict'
 
 import React from 'react'
-import { useLoginWithCode } from '@/auth/adapter/inbound/hook/UseLoginWithCode'
+import { useLoginWithCode } from '@/auth/controller/event/UseLoginWithCode'
 
 export default function Page (): React.ReactElement {
   useLoginWithCode()
