+++ src/auth/infra/atom/Session.tsx
@@ -1,0 +1,4 @@
+import { atom } from 'jotai/index'
+import type { Session } from '@/auth/application/domain/Session'
+
+export const sessionAtom = atom<Session | null>(null)
